+++ src/test/java/de/wasenweg/alfred/integration/CoverIntegrationTest.java
@@ -84,7 +84,7 @@ public class CoverIntegrationTest {
   public void deletesFirstPage() throws Exception {
     // Given
     Comic comic = this.comicRepository.findAll().get(0);
-    assertEquals(comic.getPageCount(), (short) 3);
+    assertEquals(comic.getPageCount(), 3);
     assertIterableEquals(
         comic.getFiles(), Arrays.asList("/1.png", "/2.png", "/3.png", "/ComicInfo.xml"));
 
@@ -99,7 +99,7 @@ public class CoverIntegrationTest {
 
     // Then
     comic = this.comicRepository.findAll().get(0);
-    assertEquals(comic.getPageCount(), (short) 2);
+    assertEquals(comic.getPageCount(), 2);
     assertIterableEquals(
         comic.getFiles(), Arrays.asList("/2.png", "/3.png", "/ComicInfo.xml"));
   }
@@ -109,7 +109,7 @@ public class CoverIntegrationTest {
   public void deletesLastPage() throws Exception {
     // Given
     Comic comic = this.comicRepository.findAll().get(0);
-    assertEquals(comic.getPageCount(), (short) 3);
+    assertEquals(comic.getPageCount(), 3);
     assertIterableEquals(
         comic.getFiles(), Arrays.asList("/1.png", "/2.png", "/3.png", "/ComicInfo.xml"));
 
@@ -124,7 +124,7 @@ public class CoverIntegrationTest {
 
     // Then
     comic = this.comicRepository.findAll().get(0);
-    assertEquals(comic.getPageCount(), (short) 2);
+    assertEquals(comic.getPageCount(), 2);
     assertIterableEquals(
         comic.getFiles(), Arrays.asList("/1.png", "/2.png", "/ComicInfo.xml"));
   }
