+++ src/test/java/de/wasenweg/alfred/unit/ComicConstructionTest.java
@@ -11,7 +11,7 @@ public class ComicConstructionTest {
   @Test
   public void constructWithNew() throws Exception {
     final Comic comic = new Comic();
-    assertThat(comic.getCurrentPage()).isEqualTo((short) 0);
+    assertThat(comic.getCurrentPage()).isEqualTo(0);
     assertThat(comic.isRead()).isEqualTo(false);
   }
 
@@ -25,11 +25,11 @@ public class ComicConstructionTest {
         .volume("")
         .number("")
         .position("")
-        .year((short) 2000)
-        .month((short) 5)
+        .year(2000)
+        .month(5)
         .publisher("")
         .build();
-    assertThat(comic.getCurrentPage()).isEqualTo((short) 0);
+    assertThat(comic.getCurrentPage()).isEqualTo(0);
     assertThat(comic.isRead()).isEqualTo(false);
   }
